+++ abipy/dfpt/qha.py
@@ -9,9 +9,10 @@ from monty.collections import dict2namedtuple
 from monty.functools import lazy_property
 from pymatgen.analysis.eos import EOS
 from abipy.core.func1d import Function1D
-from abipy.tools.plotting import add_fig_kwargs, get_ax_fig_plt
+from abipy.tools.plotting import add_fig_kwargs, get_ax_fig_plt, get_axarray_fig_plt
 from abipy.electrons.gsr import GsrFile
-from abipy.dfpt.phonons import PhononBandsPlotter, PhononDos
+from abipy.dfpt.ddb import DdbFile
+from abipy.dfpt.phonons import PhononBandsPlotter, PhononDos, PhdosFile
 from abipy.dfpt.gruneisen import GrunsNcFile
 
 
@@ -121,7 +122,7 @@ class AbstractQHA(metaclass=abc.ABCMeta):
 
     def set_eos(self, eos_name):
         """
-        Updates the EOS used for the fit.
+        Set the EOS model used for the fit.
 
         Args:
             eos_name: string indicating the expression used to fit the energies. See pymatgen.analysis.eos.EOS.
@@ -203,6 +204,7 @@ class AbstractQHA(metaclass=abc.ABCMeta):
         ax.plot(alpha.mesh, alpha.values, **kwargs)
         ax.set_xlabel(r'T (K)')
         ax.set_ylabel(r'$\alpha$ (K$^{-1}$)')
+        ax.grid(True)
 
         ax.set_xlim(tstart, tstop)
         ax.get_yaxis().get_major_formatter().set_powerlimits((0, 0))
@@ -212,7 +214,7 @@ class AbstractQHA(metaclass=abc.ABCMeta):
     @add_fig_kwargs
     def plot_vol_vs_t(self, tstart=0, tstop=800, num=100, ax=None, **kwargs):
         """
-        Plots the volume as a function of the temperature.
+        Plot the volume as a function of the temperature.
 
         Args:
             tstart: The starting value (in Kelvin) of the temperature mesh.
@@ -235,6 +237,7 @@ class AbstractQHA(metaclass=abc.ABCMeta):
         ax.set_xlabel('T (K)')
         ax.set_ylabel(r'V (${\AA}^3$)')
         ax.set_xlim(tstart, tstop)
+        ax.grid(True)
 
         return fig
 
@@ -284,7 +287,7 @@ class AbstractQHA(metaclass=abc.ABCMeta):
         Calculates the volume corresponding to a specific temperature.
 
         Args:
-            t: a temperature in K
+            t: temperature in K
 
         Returns:
             The volume
@@ -435,25 +438,34 @@ class QHA(AbstractQHA):
     """
 
     @classmethod
-    def from_files(cls, gsr_files_paths, phdos_files_paths):
+    def from_files(cls, gsr_paths, phdos_paths):
         """
         Creates an instance of QHA from a list of GSR files and a list of PHDOS.nc files.
         The list should have the same size and the volumes should match.
 
         Args:
-            gsr_files_paths: list of paths to GSR files.
-            phdos_files_paths: list of paths to PHDOS.nc files.
+            gsr_paths: list of paths to GSR files.
+            phdos_paths: list of paths to PHDOS.nc files.
 
         Returns: A new instance of QHA
         """
         energies = []
         structures = []
-        for gp in gsr_files_paths:
+        for gp in gsr_paths:
             with GsrFile.from_file(gp) as g:
                 energies.append(g.energy)
                 structures.append(g.structure)
 
-        doses = [PhononDos.as_phdos(dp) for dp in phdos_files_paths]
+        #doses = [PhononDos.as_phdos(dp) for dp in phdos_paths]
+
+        doses = []
+        structures_from_phdos = []
+        for path in phdos_paths:
+            with PhdosFile(path) as p:
+                doses.append(p.phdos)
+                structures_from_phdos.append(p.structure)
+
+        cls._check_volumes(structures, structures_from_phdos)
 
         return cls(structures, doses, energies)
 
@@ -468,6 +480,23 @@ class QHA(AbstractQHA):
         """
         super().__init__(structures=structures, energies=energies, eos_name=eos_name, pressure=pressure)
         self.doses = doses
+
+    @staticmethod
+    def _check_volumes(struct_list1, struct_list2):
+
+        lens = (len(struct_list1), len(struct_list2))
+        if lens[0] != lens[1]:
+            raise RuntimeError("Expecting lists with same number of structures. Got %s" % str(lens))
+        vols1 = [s.volume for s in struct_list1]
+        vols2 = [s.volume for s in struct_list2]
+        ierr = 0
+        for v1, v2 in zip(vols1, vols2):
+            if abs(v1 - v2) > 1e-3:
+                ierr += 1
+                print("Volume1: %s != Volume2: %s" % (v1, v2))
+
+        if ierr != 0:
+            raise RuntimeError("Expecting lists with same volumes!")
 
     def get_vib_free_energies(self, tstart=0, tstop=800, num=100):
         """
@@ -536,27 +565,27 @@ class QHA3PF(AbstractQHA):
     """
 
     @classmethod
-    def from_files(cls, gsr_files_paths, phdos_files_paths, ind_doses):
+    def from_files(cls, gsr_paths, phdos_paths, ind_doses):
         """
-        Creates an instance of QHA from a list of GSR files and a list PHDOS.nc files.
+        Creates an instance of QHA from a list of GSR files and a list of PHDOS.nc files.
         The list should have the same size and the volumes should match.
 
         Args:
-            gsr_files_paths: list of paths to GSR files.
-            phdos_files_paths: list of paths to three PHDOS.nc files.
+            gsr_paths: list of paths to GSR files.
+            phdos_paths: list of paths to three PHDOS.nc files.
             ind_doses: list of three values indicating, for each of the three doses, the index of the
-                corresponding gsr_file in "gsr_files_paths".
+                corresponding gsr_file in "gsr_paths".
 
         Returns: A new instance of QHA
         """
         energies = []
         structures = []
-        for gp in gsr_files_paths:
+        for gp in gsr_paths:
             with GsrFile.from_file(gp) as g:
                 energies.append(g.energy)
                 structures.append(g.structure)
 
-        doses = [PhononDos.as_phdos(dp) for dp in phdos_files_paths]
+        doses = [PhononDos.as_phdos(dp) for dp in phdos_paths]
 
         return cls(structures, doses, energies, ind_doses)
 
@@ -675,23 +704,23 @@ class QHA3P(AbstractQHA):
     """
 
     @classmethod
-    def from_files(cls, gsr_files_paths, grun_file_path, ind_doses):
+    def from_files(cls, gsr_paths, grun_file_path, ind_doses):
         """
         Creates an instance of QHA from a list of GSR files and a list PHDOS.nc files.
         The list should have the same size and the volumes should match.
 
         Args:
-            gsr_files_paths: list of paths to GSR files.
-            phdos_files_paths: list of paths to three PHDOS.nc files.
+            gsr_paths: list of paths to GSR files.
+            phdos_paths: list of paths to three PHDOS.nc files.
             ind_doses: list of three values indicating, for each of the three doses, the index of the
-                corresponding gsr_file in "gsr_files_paths".
+                corresponding gsr_file in "gsr_paths".
 
         Returns:
             A new instance of QHA
         """
         energies = []
         structures = []
-        for gp in gsr_files_paths:
+        for gp in gsr_paths:
             with GsrFile.from_file(gp) as g:
                 energies.append(g.energy)
                 structures.append(g.structure)
@@ -716,6 +745,10 @@ class QHA3P(AbstractQHA):
         self.grun = gruns
         self.ind_grun = ind_grun
         self._ind_energy_only = [i for i in range(len(structures)) if i not in ind_grun]
+
+    def close(self):
+        """Close files."""
+        self.grun.close()
 
     def get_thermodynamic_properties(self, tstart=0, tstop=800, num=100):
         """
@@ -874,3 +907,124 @@ def get_entropy(w, weights, t):
     ind = np.where(w >= 0)
 
     return np.dot(weights[ind[0]], s[ind]).sum()
+
+
+class AbstractQmeshAnalyzer(metaclass=abc.ABCMeta):
+    """
+    Abstract class for the analysis of the convergence wrt to the q-mesh used to compute the
+    phonon DOS. Relies on abstract methods implemented in AbstractQHA.
+    """
+
+    fontsize = 8
+    colormap = "viridis"
+
+    def _consistency_check(self):
+        if not hasattr(self, "qha_list") or not self.qha_list:
+            raise RuntimeError("Please call the run method to compute the QHA!")
+
+    def set_eos(self, eos_name):
+        """
+        Set the EOS model used for the fit.
+
+        Args:
+            eos_name: string indicating the expression used to fit the energies. See pymatgen.analysis.eos.EOS.
+        """
+        self._consistency_check()
+        for qha in self.qha_list:
+            qha.set_eos(eos_name)
+
+    @add_fig_kwargs
+    def plot_energies(self, **kwargs):
+        """
+        Plots the energies as a function of volume at different temperatures.
+        kwargs are propagated to the analogous method of QHA.
+        """
+        self._consistency_check()
+        # Build grid of plots.
+        ax_list, fig, plt = get_axarray_fig_plt(None, nrows=self.num_qmeshes, ncols=1,
+                                                sharex=True, sharey=True, squeeze=False)
+        ax_list = ax_list.ravel()
+
+        for qha, nqsm, ax in zip(self.qha_list, self.nqsmall_list, ax_list):
+            qha.plot_energies(ax=ax, show=False, **kwargs)
+            ax.set_title("nqsmall: %d" % nqsm, fontsize=self.fontsize)
+        return fig
+
+    @add_fig_kwargs
+    def plot_thermal_expansion_coeff(self, **kwargs):
+        """
+        Plots the thermal expansion coefficient as a function of the temperature.
+        kwargs are propagated to the analogous method of QHA.
+        """
+        self._consistency_check()
+        ax, fig, plt = get_ax_fig_plt(None)
+        cmap = plt.get_cmap(self.colormap)
+        for iq, (qha, nqsm) in enumerate(zip(self.qha_list, self.nqsmall_list)):
+            qha.plot_thermal_expansion_coeff(ax=ax,
+                                             color=cmap(float(iq) / self.num_qmeshes),
+                                             label="nqsmall: %s" % nqsm, show=False, **kwargs)
+        ax.legend(loc="best", fontsize=self.fontsize, shadow=True)
+        return fig
+
+    @add_fig_kwargs
+    def plot_vol_vs_t(self, **kwargs):
+        """
+        Plot the volume as a function of the temperature.
+        kwargs are propagated to the analogous method of QHA.
+        """
+        self._consistency_check()
+        ax, fig, plt = get_ax_fig_plt(None)
+        cmap = plt.get_cmap(self.colormap)
+        for iq, (qha, nqsm) in enumerate(zip(self.qha_list, self.nqsmall_list)):
+            qha.plot_vol_vs_t(ax=ax,
+                              color=cmap(float(iq) / self.num_qmeshes),
+                              label="nqsmall: %s" % nqsm, show=False, **kwargs)
+        ax.legend(loc="best", fontsize=self.fontsize, shadow=True)
+        return fig
+
+
+class QHAQmeshAnalyzer(AbstractQmeshAnalyzer):
+
+    def __init__(self, gsr_paths, ddb_paths):
+        """
+        Creates an instance of QHA from a list of GSR files and a list of PHDOS.nc files.
+        The list should have the same size and the volumes should match.
+
+        Args:
+            gsr_paths: list of paths to GSR files.
+            phdos_paths: list of paths to PHDOS.nc files.
+        """
+        self.gsr_paths = gsr_paths
+        self.ddb_paths = ddb_paths
+
+    def run_qlist(self, nqsmall_list, **kwargs):
+        """
+        """
+        self.qha_list = []
+        self.nqsmall_list = nqsmall_list
+
+        ddb_list = [DdbFile(p) for p in self.ddb_paths]
+        for nqsmall in nqsmall_list:
+            phdos_paths = []
+            for ddb in ddb_list:
+                phbst_file, phdos_file = ddb.anaget_phbst_and_phdos_files(
+                    nqsmall=nqsmall, qppa=None, ndivsm=1, line_density=None, asr=2, chneut=1, dipdip=1,
+                    dos_method="tetra", lo_to_splitting="automatic", ngqpt=None, qptbounds=None,
+                    anaddb_kwargs=None, verbose=0, spell_check=True,
+                    mpi_procs=1, workdir=None, manager=None)
+
+                #ph_dos = PhononDos.as_phdos(phdos_file)
+                phdos_paths.append(phdos_file.filepath)
+                #qptrlatt = phdos_file.reader.read_value("qptrlatt").T
+                #shiftq = phdos_file.reader.read_value("shiftq")
+
+                phbst_file.close()
+                phdos_file.close()
+
+            qha = QHA.from_files(self.gsr_paths, phdos_paths)
+            self.qha_list.append(qha)
+            #self.ngqpt_list.append()
+
+        #self.ngppt_list = np.reshape(ngqpt_list, (-1, 3))
+        self.num_qmeshes = len(self.nqsmall_list)
+        for ddb in ddb_list: ddb.close()