+++ abipy/dfpt/tests/test_qha.py
@@ -3,7 +3,7 @@ import os
 import warnings
 import abipy.data as abidata
 
-from abipy.dfpt.qha import QHA, QHA3PF, QHA3P
+from abipy.dfpt.qha import QHA, QHA3PF, QHA3P, QHAQmeshAnalyzer
 from abipy.dfpt.phonons import PhononBands
 from abipy.core.testing import AbipyTest
 
@@ -16,11 +16,12 @@ class QhaTest(AbipyTest):
         dirpath = os.path.join(abidata.dirpath, "refs", "si_qha")
         cls.gsr_paths = [os.path.join(dirpath, "mp-149_{:+d}_GSR.nc".format(s)) for s in cls.strains]
         cls.dos_paths = [os.path.join(dirpath, "mp-149_{:+d}_PHDOS.nc".format(s)) for s in cls.strains]
+        cls.ddb_paths = [os.path.join(dirpath, "mp-149_{:+d}_DDB".format(s)) for s in cls.strains]
         cls.phbs_list = [PhononBands.from_file(os.path.join(dirpath, "mp-149_{:+d}_PHBST.nc".format(s))) for s in
                          cls.strains[2:4]]
 
     def test_qha(self):
-        """Base tests for QHA"""
+        """Testing QHA"""
         qha = QHA.from_files(self.gsr_paths, self.dos_paths)
 
         self.assertEqual(qha.nvols, len(self.strains))
@@ -53,7 +54,13 @@ class QhaTest(AbipyTest):
             # fake temperatures to test the plotting function.
             assert qha.plot_phbs(self.phbs_list, temperatures=[10, 20], show=False)
 
+        with self.assertRaises(RuntimeError):
+            QHA.from_files(self.gsr_paths[0:2], self.dos_paths[2:0:-1])
+        with self.assertRaises(RuntimeError):
+            QHA.from_files(self.gsr_paths[0:2], self.dos_paths[0:1])
+
     def test_phonopy_object(self):
+        """Testing QHA phonopy object."""
         self.skip_if_not_phonopy()
 
         qha = QHA.from_files(self.gsr_paths, self.dos_paths)
@@ -65,6 +72,21 @@ class QhaTest(AbipyTest):
         if self.has_matplotlib():
             qha_ph.plot_thermal_expansion()
 
+    def test_qha_qmesh_analyzer(self):
+        """Testing QHAQmeshAnalyzer."""
+        qhana = QHAQmeshAnalyzer(self.gsr_paths[1:5], self.ddb_paths[1:5])
+        with self.assertRaises(RuntimeError):
+            qhana.set_eos("murnaghan")
+        with self.assertRaises(RuntimeError):
+            assert qhana.plot_energies(title="Energies as a function of V for different T", show=False)
+
+        qhana.run_qlist([2,])
+        qhana.set_eos("birch_murnaghan")
+        if self.has_matplotlib():
+            assert qhana.plot_energies(title="Energies as a function of V for different T", show=False)
+            assert qhana.plot_thermal_expansion_coeff(title="Thermal expansion as a function of T", show=False)
+            assert qhana.plot_vol_vs_t(title="Volume as a function of T", show=False)
+
 
 class Qha3pfTest(AbipyTest):
 
@@ -75,8 +97,8 @@ class Qha3pfTest(AbipyTest):
         cls.gsr_paths = [os.path.join(path, "mp-149_{:+d}_GSR.nc".format(s)) for s in cls.strains]
         cls.dos_paths = [os.path.join(path, "mp-149_{:+d}_PHDOS.nc".format(s)) for s in cls.strains]
 
-    def test_qha(self):
-        """Base tests for QHA3PF"""
+    def test_qha3pf(self):
+        """Testing QHA3PF"""
 
         qha = QHA3PF.from_files(self.gsr_paths, self.dos_paths[1:4], ind_doses=[1, 2, 3])
 
@@ -130,7 +152,7 @@ class Qha3pTest(AbipyTest):
     def tearDown(self):
         warnings.resetwarnings()
 
-    def test_qha(self):
+    def test_qha3p(self):
         """Base tests for QHA3PF"""
         qha = QHA3P.from_files(self.gsr_paths, self.gruns_path, ind_doses=[1, 2, 3])
 
@@ -169,3 +191,4 @@ class Qha3pTest(AbipyTest):
         qha_ph.run()
         if self.has_matplotlib():
             qha_ph.plot_thermal_expansion()
+        qha.close()